requirements.txt
fastapi
uvicorn[standard]
neomodel
python-dotenv
pytest
requests
